Updated link for the header and main

+++ src/main/java/com/sos/signal/counselingappointment/CounselingAppointmentRepository.java
@@ -4,6 +4,6 @@ package com.sos.signal.counselingappointment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 // Repository interface for database operations
-public interface CounselingAppointmentRepository extends JpaRepository<CounselingAppointment, Long> {
+public interface CounselingAppointmentRepository extends JpaRepository<CounselingAppointment, Integer> {
     // Add any custom methods if needed
 }
