+++ src/main/java/com/sos/signal/policecomplaint/repository/PoliceComplaintRepository.java
@@ -4,6 +4,6 @@ import com.sos.signal.policecomplaint.dto.PoliceComplaint;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 // Repository interface for database operations
-public interface PoliceComplaintRepository extends JpaRepository<PoliceComplaint, Long> {
+public interface PoliceComplaintRepository extends JpaRepository<PoliceComplaint, Integer> {
     // Add any custom methods if needed
 }
